# Optimize OTA loop to avoid unnecessary stack allocations