# What does this implement/fix?

This PR optimizes the ESPHome OTA component's loop() method to avoid expensive stack allocations when no OTA update is in progress. 

Previously, the `handle_()` method was called ~6000 times per minute, allocating 1KB+ of stack space even when no client was connected. Runtime statistics showed the OTA component consuming ~2.8 seconds per minute (4.75% of CPU time) just checking for non-existent connections.

The stack allocations that are now skipped when OTA is idle:
```cpp
void ESPHomeOTAComponent::handle_() {
  ota::OTAResponseTypes error_code = ota::OTA_RESPONSE_ERROR_UNKNOWN;
  bool update_started = false;
  size_t total = 0;
  uint32_t last_progress = 0;
  uint8_t buf[1024];  // 1KB stack allocation
  char *sbuf = reinterpret_cast<char *>(buf);
  size_t ota_size;
  uint8_t ota_features;
  std::unique_ptr<ota::OTABackend> backend;
  (void) ota_features;
#if USE_OTA_VERSION == 2
  size_t size_acknowledged = 0;
#endif
  // ... rest of function
}
```

The optimization adds a simple check in `loop()` to only call `handle_()` when:
- A client is already connected, or
- The server socket has incoming connections ready

This reduces idle loop overhead from 0.42ms to near zero when OTA is not active.

## Performance Results

Runtime statistics from ESP32 before and after optimization:

**Before optimization:**
```
esphome.ota: count=6793, avg=0.42ms, max=1ms, total=2855ms
```

**After optimization:**
```
esphome.ota: count=7033, avg=0.08ms, max=1ms, total=592ms
```

- **80% reduction** in average loop time (0.42ms → 0.08ms)
- **79% reduction** in total CPU time (2855ms → 592ms over ~60 seconds)
- Handles more iterations with less CPU usage

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):** N/A

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):** N/A

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [x] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
# No configuration changes required
ota:
  - platform: esphome
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).